re organized tests
