+++ test/python/test_aes_cbc.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 import logging
 
-from aes_modes import aes_cbc, _aes_cbc_encrypt, _aes_cbc_decrypt
+from src.python.aes_modes import aes_cbc, _aes_cbc_encrypt, _aes_cbc_decrypt
 
 logger = logging.getLogger(__name__)
