+++ test/scripts/test_rsp_parser.py
@@ -1,0 +1,15 @@
+import pytest
+from pathlib import Path 
+from scripts.aes import aes_cbc
+from scripts.rsp_parser import parse_kat_file
+
+kat_cbc_files = Path("KAT_AES").glob("CBC*.rsp")
+@pytest.mark.parametrize("kat_files", kat_cbc_files, ids=lambda p: p.name)
+def test_kat_aes_cbc(kat_files):
+    kat_list = parse_kat_file(kat_files)
+    for (op, count), value in kat_list.items():
+        aes_cbc(op, value.key, value.iv, value.plaintext, value.ciphertext)
+
+    
+
+
